Refactor password validator. Add tests
