+++ README.md
@@ -1,1 +1,11 @@
-# homework
+# Homework
+
+## Build & test
+```$xslt
+mvn clean install
+```
+
+## Run
+```$xslt
+mvn exec:java
+```
